Update EFHST version and add semver vuln' patch
New EFHST version includes the changes for the webpack base config to provide the fallbacks for query string, meaning it is no longer required in this package.

Currently there is a compile time vulnerability with semver. An override was added to force it to the newest version as it should be stable even for old packages. Semver provides functionality to check NPM versions and is relatively stable otherwise.