Add support for page fragments to Jakarta site (#1686)
* Add support for page fragments to Jakarta site

Using new version of Hugo theme, a new templates section was added that allows the sites base page template to be retrieved programatically. These pages do not contain any page content, and will need to be added manually to be useful.

* Update netlify build to use NPM + v110 Hugo

* Fix missing polyfills in webpack build

* Remove unwanted build file from npm

* Update EFHST version and add semver vuln' patch

New EFHST version includes the changes for the webpack base config to provide the fallbacks for query string, meaning it is no longer required in this package.

Currently there is a compile time vulnerability with semver. An override was added to force it to the newest version as it should be stable even for old packages. Semver provides functionality to check NPM versions and is relatively stable otherwise.

* Set fixed version for EFHST